últimas modificaciones

+++ src/detalleCompraIngrediente/entity/DetalleCompraIngrediente.java
@@ -14,41 +14,50 @@ public class DetalleCompraIngrediente {
 		this.totalCompra = totalCompra;
 		
 	}
+	
 	public int getCodDetalle() {
 		return codDetalle;
 	}
+	
 	public void setCodDetalle(int codDetalle) {
 		this.codDetalle = codDetalle;
 	}
+	
 	public int getCodCompraIng() {
 		return codCompraIng;
 	}
+	
 	public void setCodCompraIng(int codCompraIng) {
 		this.codCompraIng = codCompraIng;
 	}
+	
 	public int getCodIngrediente() {
 		return codIngrediente;
 	}
+	
 	public void setCodIngrediente(int codIngrediente) {
 		this.codIngrediente = codIngrediente;
 	}
+	
 	public int getCantidad() {
 		return cantidad;
 	}
+	
 	public void setCantidad(int cantidad) {
 		this.cantidad = cantidad;
 	}
+	
 	public double getTotalCompra() {
 		return totalCompra;
 	}
-	public void setTotalCompra(int totalCompra) {
+	
+	public void setTotalCompra(double totalCompra) {
 		this.totalCompra = totalCompra;
 	}
 	
 	@Override
 	public String toString() {
 		return "DetalleVenta [codDetalle=" + codDetalle + ", codCompraIng=" + codCompraIng
-				+ ", codIngrediente=" + codIngrediente
-				+ ", cantidad=" + cantidad + ", totalCompra=" + totalCompra + "]";
+				+ ", codIngrediente=" + codIngrediente + ", cantidad=" + cantidad + ", totalCompra=" + totalCompra + "]";
 	}
 }
